+++ source/includes/code-examples/insert-many/InsertManyAsync.cs
@@ -4,12 +4,12 @@ using MongoDB.Bson.Serialization.Conventions;
 using MongoDB.Driver;
 using static System.Console;
 
-namespace CsharpExamples.UsageExamples;
+namespace CsharpExamples.UsageExamples.InsertMany;
 
 public class InsertManyAsync
 {
     private static IMongoCollection<Restaurant> _restaurantsCollection;
-    private static string _mongoConnectionString = "<Your MongoDB URI>";
+    private const string _mongoConnectionString = "<Your MongoDB URI>";
 
     public static void Main(string[] args)
     {
@@ -28,7 +28,7 @@ public class InsertManyAsync
         WriteLine("Inserting documents...");
         InsertManyRestaurantsAsync();
 
-        // find and print newly inserted document
+        // Find and print newly inserted document
         foundRestaurants = _restaurantsCollection.Find(filter).ToList();
 
         WriteLine($"Number of restaurants inserted: {foundRestaurants.Count}");
@@ -38,7 +38,7 @@ public class InsertManyAsync
 
     private static async void InsertManyRestaurantsAsync()
     {
-        // delete sample document if already exists
+        // Delete sample document if already exists
         Cleanup();
 
         // start-insert-many
@@ -93,21 +93,3 @@ public class InsertManyAsync
         _restaurantsCollection.DeleteMany(filter);
     }
 }
-
-// start-model
-public class Restaurant
-{
-    public ObjectId Id { get; set; }
-
-    public string Name { get; set; }
-
-    [BsonElement("restaurant_id")]
-    public string RestaurantId { get; set; }
-
-    public string Cuisine { get; set; }
-
-    public object Address { get; set; }
-
-    public string Borough { get; set; }
-}
-// end-model
